+++ xonsh/xonfig.py
@@ -324,6 +324,8 @@ def _tok_colors(cmap, cols):
 
 def _colors(ns):
     cols, _ = shutil.get_terminal_size()
+    if tools.ON_WINDOWS:
+        cols -= 1
     cmap = tools.color_style()
     akey = next(iter(cmap))
     if isinstance(akey, str):